Edited Option 2 to remove unnecessary codes
Remove unnecessary codes in option 2 of SchoolSystem and make changes in Course.java so that option 2 can call on it. (*Ignore Course.java codes for now as I am working to clean it up and structured it in a nicer way. It will be confusing as for now to read. Codes for validating and printing are in but still working on it make sure it function properly.)